# perf: replace 'input_to_float8' with sgl_per_tensor_quant_fp8 kernel to get 6% throughtput improvement